+++ flutter_reach_five/test/helpers/adapt_errors_test.dart
@@ -158,5 +158,20 @@ void main() {
 
       expect(error, isA<AccountBlockedAfterMultipleLoginAttemptsException>());
     });
+
+    test('$TooManyAttemptsException', () async {
+      Object? error;
+
+      try {
+        adaptErrors(
+          error: const TooManyAttemptsExceptionInterface(),
+          stackTrace: StackTrace.fromString('test'),
+        );
+      } catch (e) {
+        error = e;
+      }
+
+      expect(error, isA<TooManyAttemptsException>());
+    });
   });
 }