+++ flutter_reach_five_platform_interface/lib/src/errors.dart
@@ -11,7 +11,9 @@ final errorCodesInterface = ErrorCodesInterface(
   userCancelledOrClosedTheWebFlow: 'user_cancelled_or_closed_the_web_flow',
   socialAccountEmailAlreadyInUse: 'social_account_email_already_in_use',
   unauthorizedRefreshToken: 'unauthorized_refresh_token',
-  accountBlockedAfterMultipleLoginAttempts: 'too_many_login_attemps',
+  accountBlockedAfterMultipleLoginAttempts:
+      'account_blocked_too_many_login_attempts',
+  tooManyAttempts: 'rate_limiter_too_many_attempts',
 );
 
 /// {@macro flutter_reach_five.errors.email_already_in_use}
@@ -39,13 +41,21 @@ class SocialAccountEmailAlreadyInUseExceptionInterface implements Exception {}
 /// {@macro flutter_reach_five.errors.unauthorized_refresh_token}
 class UnauthorizedRefreshTokenExceptionInterface implements Exception {}
 
-/// {@macro flutter_reach_five.errors.too_many_login_attemps}
+/// {@macro flutter_reach_five.errors.account_blocked_too_many_login_attempts}
 class AccountBlockedAfterMultipleLoginAttemptsExceptionInterface
     extends FormatException {
   // coverage:ignore-start
-  /// {@macro flutter_reach_five.errors.too_many_login_attemps}
+  /// {@macro flutter_reach_five.errors.account_blocked_too_many_login_attempts}
   const AccountBlockedAfterMultipleLoginAttemptsExceptionInterface([
     super.message = '',
   ]);
+  // coverage:ignore-end
+}
+
+/// {@macro flutter_reach_five.errors.too_many_attempts}
+class TooManyAttemptsExceptionInterface extends FormatException {
+  // coverage:ignore-start
+  /// {@macro flutter_reach_five.errors.too_many_attempts}
+  const TooManyAttemptsExceptionInterface([super.message = '']);
   // coverage:ignore-end
 }