+++ flutter_reach_five_platform_interface/lib/src/flutter_reach_five_platform.dart
@@ -161,6 +161,7 @@ abstract class FlutterReachFivePlatform extends PlatformInterface {
       email: email,
       phoneNumber: phoneNumber,
       redirectUrl: redirectUrl,
+      errorCodes: errorCodesInterface,
     ),
   );